node-main/deps/v8/test/inspector/debugger/wasm-global-names-expected.txt

Test wasm global names

Running test: test
Waiting for wasm script to be parsed.
Setting breakpoint in wasm.
Running main.
Paused in debugger.
   globals: {"$module_name.imported_global", "$exported_global", "$global2"}
